+++ dd-java-agent/agent-tooling/src/main/java/datadog/trace/agent/tooling/muzzle/MuzzleGradlePlugin.java
@@ -1,36 +1,22 @@
 package datadog.trace.agent.tooling.muzzle;
 
-import datadog.trace.agent.tooling.HelperInjector;
 import datadog.trace.agent.tooling.Instrumenter;
-import java.lang.reflect.Method;
-import java.util.List;
-import java.util.ServiceLoader;
+import net.bytebuddy.asm.AsmVisitorWrapper;
 import net.bytebuddy.build.Plugin;
+import net.bytebuddy.description.field.FieldDescription;
+import net.bytebuddy.description.field.FieldList;
+import net.bytebuddy.description.method.MethodList;
 import net.bytebuddy.description.type.TypeDefinition;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.dynamic.DynamicType.Builder;
+import net.bytebuddy.implementation.Implementation;
+import net.bytebuddy.jar.asm.ClassVisitor;
+import net.bytebuddy.jar.asm.MethodVisitor;
+import net.bytebuddy.jar.asm.Opcodes;
+import net.bytebuddy.pool.TypePool;
 
+/** Bytebuddy gradle plugin which creates muzzle-references at compile time. */
 public class MuzzleGradlePlugin implements Plugin {
-  // TODO:
-  // - Additional references to check
-  //   - Fields
-  //   - methods
-  //     - visit annotations
-  //     - visit parameter types
-  //     - visit method instructions
-  //     - method invoke type
-  //   - access flags (including implicit package-private)
-  //   - supertypes
-  // - Misc
-  //   - Also match interfaces which extend Instrumenter
-  //   - Expose config instead of hardcoding datadog namespace (or reconfigure classpath)
-  //   - Run muzzle in matching phase (may require a rewrite of the instrumentation api)
-  //   - Documentation
-  //   - Fix TraceConfigInstrumentation
-  //   - assert no muzzle field defined in instrumentation
-  //   - make getMuzzle final in default and remove in gradle plugin
-  //   - pull muzzle field + method names into static constants
-
   private static final TypeDescription DefaultInstrumenterTypeDesc =
       new TypeDescription.ForLoadedType(Instrumenter.Default.class);
 
@@ -38,7 +24,7 @@ public class MuzzleGradlePlugin implements Plugin {
   public boolean matches(final TypeDescription target) {
     // AutoService annotation is not retained at runtime. Check for Instrumenter.Default supertype
     boolean isInstrumenter = false;
-    TypeDefinition instrumenter = target;
+    TypeDefinition instrumenter = null == target ? null : target.getSuperClass();
     while (instrumenter != null) {
       if (instrumenter.equals(DefaultInstrumenterTypeDesc)) {
         isInstrumenter = true;
@@ -51,15 +37,10 @@ public class MuzzleGradlePlugin implements Plugin {
 
   @Override
   public Builder<?> apply(Builder<?> builder, TypeDescription typeDescription) {
-    if (typeDescription.equals(DefaultInstrumenterTypeDesc)) {
-      // FIXME
-      System.out.println("~~~~FIXME: remove final modifier on Default: " + typeDescription);
-      return builder;
-    } else {
-      return builder.visit(new MuzzleVisitor());
-    }
+    return builder.visit(new MuzzleVisitor());
   }
 
+  /** Compile-time Optimization used by gradle buildscripts. */
   public static class NoOp implements Plugin {
     @Override
     public boolean matches(final TypeDescription target) {
@@ -72,53 +53,56 @@ public class MuzzleGradlePlugin implements Plugin {
     }
   }
 
-  public static void assertAllInstrumentationIsMuzzled(ClassLoader cl) throws Exception {
-    for (Instrumenter instrumenter :
-        ServiceLoader.load(Instrumenter.class, MuzzleGradlePlugin.class.getClassLoader())) {
-      if (instrumenter.getClass().getName().endsWith("TraceConfigInstrumentation")) {
-        // TraceConfigInstrumentation doesn't do muzzle checks
-        // check on TracerClassInstrumentation instead
-        instrumenter =
-            (Instrumenter)
-                MuzzleGradlePlugin.class
-                    .getClassLoader()
-                    .loadClass(instrumenter.getClass().getName() + "$TracerClassInstrumentation")
-                    .getDeclaredConstructor()
-                    .newInstance();
+  private static class RemoveFinalFlagVisitor implements AsmVisitorWrapper {
+    final String methodName;
+
+    public RemoveFinalFlagVisitor(String methodName) {
+      this.methodName = methodName;
+    }
+
+    @Override
+    public int mergeWriter(int flags) {
+      return flags;
+    }
+
+    @Override
+    public int mergeReader(int flags) {
+      return flags;
+    }
+
+    @Override
+    public ClassVisitor wrap(
+        TypeDescription instrumentedType,
+        ClassVisitor classVisitor,
+        Implementation.Context implementationContext,
+        TypePool typePool,
+        FieldList<FieldDescription.InDefinedShape> fields,
+        MethodList<?> methods,
+        int writerFlags,
+        int readerFlags) {
+      return new Visitor(classVisitor);
+    }
+
+    private class Visitor extends ClassVisitor {
+      public Visitor(ClassVisitor cv) {
+        super(Opcodes.ASM6, cv);
       }
-      Method m = null;
-      try {
-        m = instrumenter.getClass().getDeclaredMethod("getInstrumentationMuzzle");
-        m.setAccessible(true);
-        ReferenceMatcher muzzle = (ReferenceMatcher) m.invoke(instrumenter);
-        List<Reference.Mismatch> mismatches = muzzle.getMismatchedReferenceSources(cl);
-        if (mismatches.size() > 0) {
-          System.err.println(
-              "FAILED MUZZLE VALIDATION: " + instrumenter.getClass().getName() + " mismatches:");
-          for (Reference.Mismatch mismatch : mismatches) {
-            System.err.println("-- " + mismatch);
-          }
-          throw new RuntimeException("Instrumentation failed Muzzle validation");
+
+      @Override
+      public MethodVisitor visitMethod(
+          final int access,
+          final String name,
+          final String descriptor,
+          final String signature,
+          final String[] exceptions) {
+        MethodVisitor methodVisitor =
+            super.visitMethod(access, name, descriptor, signature, exceptions);
+        if (name.equals(methodName) && (access & Opcodes.ACC_FINAL) != 0) {
+          return super.visitMethod(
+              access ^ Opcodes.ACC_FINAL, name, descriptor, signature, exceptions);
+        } else {
+          return super.visitMethod(access, name, descriptor, signature, exceptions);
         }
-      } finally {
-        if (null != m) {
-          m.setAccessible(false);
-        }
-      }
-      try {
-        // Ratpack injects the scope manager as a helper.
-        // This is likely a bug, but we'll grandfather it out of the helper checks for now.
-        if (!instrumenter.getClass().getName().contains("Ratpack")) {
-          // verify helper injector works
-          final String[] helperClassNames = instrumenter.helperClassNames();
-          if (helperClassNames.length > 0) {
-            new HelperInjector(helperClassNames).transform(null, null, cl, null);
-          }
-        }
-      } catch (Exception e) {
-        System.err.println(
-            "FAILED HELPER INJECTION. Are Helpers being injected in the correct order?");
-        throw e;
       }
     }
   }
